+++ src/main/java/com/bernardomg/darksouls/explorer/item/model/ItemMerchantSource.java
@@ -3,6 +3,6 @@ package com.bernardomg.darksouls.explorer.item.model;
 
 public interface ItemMerchantSource extends ItemSource {
 
-    public Float getCost();
+    public Double getCost();
 
 }
